cleanup(test): rename e2e to functional
