+++ functional/servers/openai/types.go
